requirements.txt
crewai
crewai_tools
langchain
langchain_google_genai
langchain_groq
streamlit
python-dotenv
SerperDevTool
openai
